servidores/requirements.txt
appdata==2.1.2
asgiref==3.5.2
click==8.0.3
colored==1.4.3
Django
django-heroku==0.3.1
djangorestframework==3.13.1
mysql-connector==2.2.9
mysql-connector-python==8.0.11
mysqlclient==2.1.1
protobuf==3.20.1
PyMySQL==1.0.2
pytz==2022.2.1
psycopg2==2.8.5
psycopg2-binary==2.9.3
sqlparse==0.4.2
to-requirements.txt==1.3.0
webdriver-manager==3.8.1
webdrivermanager==0.10.0
websockets==10.3
whitenoise==5.2.0
gunicorn==20.0.4
